Moved the scheduling functions to a utils file
Added
- schedule_utils.py for the scheduling algorithm's helper functions

Fixed
- Import errors
- The scheduling algorithm wasn't following the SOLID principle
